Remove remaining usage of height, as words is the key.
